+++ benchmark/brotli_benchmark.dart
@@ -8,6 +8,7 @@ import 'package:benchmark_harness/benchmark_harness.dart';
 import 'package:collection/collection.dart';
 import 'package:es_compression/brotli.dart';
 import 'package:es_compression/framework.dart';
+import 'package:meta/meta.dart';
 
 import 'utils/benchmark_utils.dart';
 
@@ -99,31 +100,39 @@ class BrotliData {
 /// This means the final decoded bytes should match the original input.
 /// Verify this and report success (0) if good, failure (-1) if the bytes
 /// don't match.
-void main() {
-  const dataLength = 50 * 1024 * 1024;
-  print('generating $dataLength bytes of random data');
-  final bytes = generateRandomBytes(dataLength);
-  final emitter = CodecPerformanceEmitter(bytes.length);
+Future<void> main() async {
+  const dataLength = 50 * 1024 * 1024; // 50 MB;
+  exitCode = await runBrotliBenchmark(dataLength);
+}
 
-  print('Brotli encode/decode ${bytes.length} bytes of random data.');
-  var data = BrotliData(bytes);
-  BrotliEncodeBenchmark(data, emitter: emitter).report();
-  print(
-      'compression ratio: ${compressionRatio(bytes.length, data.bytes.length)}');
-  BrotliDecodeBenchmark(data, emitter: emitter).report();
-  var bytesMatch = const ListEquality<int>().equals(bytes, data.bytes);
-  if (bytesMatch != true) exit(-1);
+/// Brotli Benchmark which answers 0 on success, -1 on error
+@visibleForTesting
+Future<int> runBrotliBenchmark(int dataLength) async {
+  return Future(() {
+    print('generating $dataLength bytes of random data');
+    final bytes = generateRandomBytes(dataLength);
+    final emitter = CodecPerformanceEmitter(bytes.length);
 
-  print('');
-  print('generating ${bytes.length} bytes of constant data');
-  bytes.fillRange(0, bytes.length, 1);
+    print('Brotli encode/decode ${bytes.length} bytes of random data.');
+    var data = BrotliData(bytes);
+    BrotliEncodeBenchmark(data, emitter: emitter).report();
+    print(
+        'compression ratio: ${compressionRatio(bytes.length, data.bytes.length)}');
+    BrotliDecodeBenchmark(data, emitter: emitter).report();
+    var bytesMatch = const ListEquality<int>().equals(bytes, data.bytes);
+    if (bytesMatch != true) return -1;
 
-  print('Brotli encode/decode ${bytes.length} bytes of constant data.');
-  data = BrotliData(bytes);
-  BrotliEncodeBenchmark(data, emitter: emitter).report();
-  print(
-      'compression ratio: ${compressionRatio(bytes.length, data.bytes.length)}');
-  BrotliDecodeBenchmark(data, emitter: emitter).report();
-  bytesMatch = const ListEquality<int>().equals(bytes, data.bytes);
-  if (bytesMatch != true) exit(-1);
+    print('');
+    print('generating ${bytes.length} bytes of constant data');
+    bytes.fillRange(0, bytes.length, 1);
+
+    print('Brotli encode/decode ${bytes.length} bytes of constant data.');
+    data = BrotliData(bytes);
+    BrotliEncodeBenchmark(data, emitter: emitter).report();
+    print(
+        'compression ratio: ${compressionRatio(bytes.length, data.bytes.length)}');
+    BrotliDecodeBenchmark(data, emitter: emitter).report();
+    bytesMatch = const ListEquality<int>().equals(bytes, data.bytes);
+    return (bytesMatch != true) ? -1 : 0;
+  });
 }