+++ example/brotli_example.dart
@@ -6,6 +6,7 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:es_compression/brotli.dart';
+import 'package:meta/meta.dart';
 
 import 'utils/example_utils.dart';
 
@@ -16,7 +17,13 @@ const level = 0;
 ///
 /// The [exitCode] of this script is 0 if the decoded bytes match the original,
 /// otherwise the [exitCode] is -1.
-void main() {
+Future<void> main() async {
+  exitCode = await runBrotliExample();
+}
+
+/// Brotli Example which answers 0 on success, -1 on error
+@visibleForTesting
+Future<int> runBrotliExample() async {
   final randomBytes = generateRandomBytes(randomByteCount);
   final codec = BrotliCodec(level: level);
 
@@ -25,7 +32,7 @@ void main() {
 
   // One-shot encode/decode
   final encoded = codec.encode(randomBytes);
-  final decoded = codec.decode(encoded);
+  var decoded = codec.decode(encoded);
   final oneShotResult =
       verifyEquality(randomBytes, decoded, header: 'One-shot: ');
 
@@ -33,15 +40,14 @@ void main() {
   // Split the random bytes into 10 buckets
   final chunks = splitIntoChunks(randomBytes, 10);
   final randomStream = Stream.fromIterable(chunks);
-  randomStream
+  decoded = await randomStream
       .transform(codec.encoder)
       .transform(codec.decoder)
       .fold<List<int>>(<int>[], (buffer, data) {
     buffer.addAll(data);
     return buffer;
-  }).then((decoded) {
-    final streamResult =
-        verifyEquality(randomBytes, decoded, header: 'Streaming: ');
-    exitCode = (oneShotResult == true && streamResult == true) ? 0 : -1;
   });
+  final streamResult =
+      verifyEquality(randomBytes, decoded, header: 'Streaming: ');
+  return (oneShotResult == true && streamResult == true) ? 0 : -1;
 }