+++ lib/src/brotli/decoder.dart
@@ -151,7 +151,7 @@ class _BrotliDecompressFilter extends CodecFilter<Pointer<Uint8>,
   @override
   int doFinalize(CodecBuffer outputBuffer) {
     if (!_dispatcher.callBrotliDecoderIsFinished(_brotliState)) {
-      throw StateError('Failure to finish decoding');
+      throw FormatException('Failure to finish decoding');
     }
     return 0;
   }